// resources/js/types/store-profile.ts
export type StoreProfile = {
    name: string;
    logoUrl: string | null;
    logoWideUrl: string | null;
    shortDescription: string;
    contactEmail: string;
    contactPhone: string;
    address: string;
    facebookUrl: string | null;
    instagramUrl: string | null;
    tiktokUrl: string | null;
    youtubeUrl: string | null;
    currency: string;
};